+++ packages/plugin-packer/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.1.2](https://github.com/kintone/js-sdk/compare/@kintone/plugin-packer@9.1.1...@kintone/plugin-packer@9.1.2) (2026-06-03)
+
+
+### Bug Fixes
+
+* **deps:** update dependency ajv to v8.20.0 ([01b7e1b](https://github.com/kintone/js-sdk/commit/01b7e1b0a95e697756bee1ecbe868c4cd9e81448))
+
+## [9.1.1](https://github.com/kintone/js-sdk/compare/@kintone/plugin-packer@9.1.0...@kintone/plugin-packer@9.1.1) (2026-04-08)
+
+
+### Bug Fixes
+
+* pin dependency versions and remove unused core-js ([#3697](https://github.com/kintone/js-sdk/issues/3697)) ([5138e95](https://github.com/kintone/js-sdk/commit/5138e957cb2e85e054d555ab5524c3da5bf0740b))
+
 ## [9.1.0](https://github.com/kintone/js-sdk/compare/@kintone/plugin-packer@9.0.3...@kintone/plugin-packer@9.1.0) (2026-02-11)