+++ packages/webpack-plugin-kintone-plugin/CHANGELOG.md
@@ -3,6 +3,27 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.0.5](https://github.com/kintone/js-sdk/compare/@kintone/webpack-plugin-kintone-plugin@9.0.4...@kintone/webpack-plugin-kintone-plugin@9.0.5) (2026-07-23)
+
+
+### Bug Fixes
+
+* **deps:** address security vulnerabilities in transitive dependencies ([#3896](https://github.com/kintone/js-sdk/issues/3896)) ([deffbdc](https://github.com/kintone/js-sdk/commit/deffbdca78bef94cbc1db382d0f7d07ab3208208))
+
+## [9.0.4](https://github.com/kintone/js-sdk/compare/@kintone/webpack-plugin-kintone-plugin@9.0.3...@kintone/webpack-plugin-kintone-plugin@9.0.4) (2026-05-28)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @kintone/cli to v1.19.3 ([#3814](https://github.com/kintone/js-sdk/issues/3814)) ([9a4cf57](https://github.com/kintone/js-sdk/commit/9a4cf57cfc84359e544f3c965c720f114c1f7b75))
+
+## [9.0.3](https://github.com/kintone/js-sdk/compare/@kintone/webpack-plugin-kintone-plugin@9.0.2...@kintone/webpack-plugin-kintone-plugin@9.0.3) (2026-04-08)
+
+
+### Bug Fixes
+
+* pin dependency versions and remove unused core-js ([#3697](https://github.com/kintone/js-sdk/issues/3697)) ([5138e95](https://github.com/kintone/js-sdk/commit/5138e957cb2e85e054d555ab5524c3da5bf0740b))
+
 ## [9.0.2](https://github.com/kintone/js-sdk/compare/@kintone/webpack-plugin-kintone-plugin@9.0.1...@kintone/webpack-plugin-kintone-plugin@9.0.2) (2026-03-31)